fix : 되돌아간 작업 다시 실행
