+++ Assets/Script/Interface.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 05102c5057991d94d9b6e2818b90f22b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
