+++ Assets/Script/Skills.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9cbedf7107476284ba7426b8817a91aa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
