+++ Assets/Script/Skills/Skill.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 813778a86aa3da74681056a984372ff6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
